# Full three dimsension local coordinates for intersection