--- a/core/include/detray/coordinates/cartesian2.hpp
+++ b/core/include/detray/coordinates/cartesian2.hpp
@@ -53,40 +53,32 @@ struct cartesian2 final : public coordinate_base<cartesian2, transform3_t> {
 
     /// @}
 
-    /** This method transform from a point from 2D cartesian frame to a 2D
-     * cartesian point */
-    DETRAY_HOST_DEVICE
-    inline point2 operator()(const point2 &local2) const { return local2; }
-
-    /** This method transform from a point from 3D cartesian frame to a 2D
-     * cartesian point */
-    DETRAY_HOST_DEVICE
-    inline point2 operator()(const point3 &local3) const {
-
-        return {local3[0], local3[1]};
-    }
-
     /** This method transform from a point from global cartesian 3D frame to a
      * local 2D cartesian point */
     DETRAY_HOST_DEVICE
-    inline point2 global_to_local(const transform3_t &trf3, const point3 &p,
+    inline point3 global_to_local(const transform3_t &trf3, const point3 &p,
                                   const vector3 & /*d*/) const {
-        const auto local3 = trf3.point_to_local(p);
-        return this->operator()(local3);
+        return trf3.point_to_local(p);
+    }
+
+    /** This method transform from a local 2D cartesian point to a point global
+     * cartesian 3D frame*/
+    DETRAY_HOST_DEVICE inline point3 local_to_global(const transform3_t &trf3,
+                                                     const point3 &p) const {
+        return trf3.point_to_global(p);
     }
 
     /** This method transform from a local 2D cartesian point to a point global
      * cartesian 3D frame*/
     template <typename mask_t>
-    DETRAY_HOST_DEVICE inline point3 local_to_global(
+    DETRAY_HOST_DEVICE inline point3 bound_local_to_global(
         const transform3_t &trf3, const mask_t & /*mask*/, const point2 &p,
         const vector3 & /*d*/) const {
-        return trf3.point_to_global(point3{p[0], p[1], 0.f});
+
+        return this->local_to_global(trf3, {p[0], p[1], 0.f});
     }
 
-    template <typename mask_t>
     DETRAY_HOST_DEVICE inline vector3 normal(const transform3_t &trf3,
-                                             const mask_t & /*mask*/,
                                              const point3 & /*pos*/,
                                              const vector3 & /*dir*/) const {
         vector3 ret;
@@ -98,22 +90,19 @@ struct cartesian2 final : public coordinate_base<cartesian2, transform3_t> {
         return ret;
     }
 
-    template <typename mask_t>
     DETRAY_HOST_DEVICE inline rotation_matrix reference_frame(
-        const transform3_t &trf3, const mask_t & /*mask*/,
-        const point3 & /*pos*/, const vector3 & /*dir*/) const {
+        const transform3_t &trf3, const point3 & /*pos*/,
+        const vector3 & /*dir*/) const {
         return trf3.rotation();
     }
 
-    template <typename mask_t>
     DETRAY_HOST_DEVICE inline free_to_path_matrix path_derivative(
-        const transform3_t &trf3, const mask_t &mask, const point3 &pos,
-        const vector3 &dir) const {
+        const transform3_t &trf3, const point3 &pos, const vector3 &dir) const {
 
         free_to_path_matrix derivative =
             matrix_operator().template zero<1u, e_free_size>();
 
-        const vector3 normal = this->normal(trf3, mask, pos, dir);
+        const vector3 normal = this->normal(trf3, pos, dir);
 
         const vector3 pos_term = -1.f / vector::dot(normal, dir) * normal;
 
@@ -124,12 +113,11 @@ struct cartesian2 final : public coordinate_base<cartesian2, transform3_t> {
         return derivative;
     }
 
-    template <typename mask_t>
     DETRAY_HOST_DEVICE inline void set_bound_pos_to_free_pos_derivative(
         bound_to_free_matrix &bound_to_free_jacobian, const transform3_t &trf3,
-        const mask_t &mask, const point3 &pos, const vector3 &dir) const {
+        const point3 &pos, const vector3 &dir) const {
 
-        const rotation_matrix frame = reference_frame(trf3, mask, pos, dir);
+        const rotation_matrix frame = reference_frame(trf3, pos, dir);
 
         // Get d(x,y,z)/d(loc0, loc1)
         const matrix_type<3, 2> bound_pos_to_free_pos_derivative =
@@ -140,12 +128,11 @@ struct cartesian2 final : public coordinate_base<cartesian2, transform3_t> {
                                              e_free_pos0, e_bound_loc0);
     }
 
-    template <typename mask_t>
     DETRAY_HOST_DEVICE inline void set_free_pos_to_bound_pos_derivative(
         free_to_bound_matrix &free_to_bound_jacobian, const transform3_t &trf3,
-        const mask_t &mask, const point3 &pos, const vector3 &dir) const {
+        const point3 &pos, const vector3 &dir) const {
 
-        const rotation_matrix frame = reference_frame(trf3, mask, pos, dir);
+        const rotation_matrix frame = reference_frame(trf3, pos, dir);
         const rotation_matrix frameT = matrix_operator().transpose(frame);
 
         // Get d(loc0, loc1)/d(x,y,z)
@@ -157,11 +144,10 @@ struct cartesian2 final : public coordinate_base<cartesian2, transform3_t> {
                                              e_bound_loc0, e_free_pos0);
     }
 
-    template <typename mask_t>
     DETRAY_HOST_DEVICE inline void set_bound_angle_to_free_pos_derivative(
         bound_to_free_matrix & /*bound_to_free_jacobian*/,
-        const transform3_t & /*trf3*/, const mask_t & /*mask*/,
-        const point3 & /*pos*/, const vector3 & /*dir*/) const {
+        const transform3_t & /*trf3*/, const point3 & /*pos*/,
+        const vector3 & /*dir*/) const {
         // Do nothing
     }