--- a/core/include/detray/core/detector.hpp
+++ b/core/include/detray/core/detector.hpp
@@ -440,13 +440,24 @@ class detector {
     inline const bfield_type &get_bfield() const { return _bfield; }
 
     DETRAY_HOST_DEVICE
-    inline point2 global_to_local(const geometry::barcode bc, const point3 &pos,
+    inline point3 global_to_local(const geometry::barcode bc, const point3 &pos,
                                   const vector3 &dir) const {
         const auto &sf =
             *(surfaces().begin() + static_cast<std::ptrdiff_t>(bc.index()));
         const auto ret =
             _masks.template visit<detail::global_to_local<transform3>>(
                 sf.mask(), _transforms[sf.transform()], pos, dir);
+        return ret;
+    }
+
+    DETRAY_HOST_DEVICE
+    inline point3 local_to_global(const geometry::barcode bc,
+                                  const point3 &local) const {
+        const auto &sf =
+            *(surfaces().begin() + static_cast<std::ptrdiff_t>(bc.index()));
+        const auto ret =
+            _masks.template visit<detail::local_to_global<transform3>>(
+                sf.mask(), _transforms[sf.transform()], local);
         return ret;
     }