This PR removes the `intersection2D_point` - This a bit complicates the test code but I think it is better to clean up the codes not used in the actual applications 

## UPDATE

I made more changes to simplify the intersector and coordinates.

### Coordinates

So far the local coordinate has been described by two bound parameters
1. cartesian2 (x,y)
2. polar2 (r, phi)
3. cylinder2 (r*phi, z)
4. line2 (r, z)

The problem was that two bound local parameters are not enough to reconstruct the global position for some coordinates. Thus I decided to describe local position with three parameters:

1. cartesian2 (x,y,z)
2. polar2 (r, phi, z)
3. cylinder2 (r*phi, z, r)
4. line2 (r, z, phi) 

The bound track parameters take only the first two local parameters so there is no change in actual tracking algorithm.

There are three coordinate transform functions for every coordinates.  
1. `global_to_local` converts `point3` global position to `point3` local position.
2. `local_to_global` converts `point3` local position to `point3` global position.
3. `bound_local_to_global` converts `point2` bound local position to `point3` global position. This is equivalent to `local_to_global` function of the main branch.

It was also nice that I could remove `mask` input argument from the most of functions in coordinates

### Masks

The advantage of using full three parameters in coordinate is that we don't have to separate the `local_frame` and `measurement_frame`  anymore. And `cylindrical2D` also doesn't need to be defined with `cylindrical3` for radial check because radial information is in the local coordinate

### Intersector

We don't have to have if statement for `radial_check` and `square_cross_section`  for `is_inside` argument because local coordinate holds all information we need