@beomki-yeo 2023-04-26
Let me excuse before @niermann999 reasons: As this PR increases the intersection data size, the speed performance drops ~10%. (It is still not very clear if it is all from the data size increase though). But let's not worry too much about it right now. Once the local navigation is deployed, the effect from the data size will be relatively small.

```
// Before
CUDA sync propagation/256    145214249 ns    145203152 ns            5 TracksPropagated=451.34k/s
// After
CUDA sync propagation/256    133035119 ns    133032181 ns            5 TracksPropagated=492.633k/s
```